project zemn.me: move timeline into components/
