+++ project/zemn.me/next/BUILD.bazel
@@ -14,6 +14,7 @@ ts_project(
         "//project/zemn.me/bio",
         "//project/zemn.me/elements/TimeEye",
         "//project/zemn.me/next/assets/kenwood",
+        "//project/zemn.me/next/components",
         "//ts/next.js",
         "//ts/react/lang",
     ],
