+++ project/zemn.me/next/components/timeline/index.tsx
@@ -1,0 +1,2 @@
+export * from 'project/zemn.me/next/components/timeline/timeline';
+export { default } from 'project/zemn.me/next/components/timeline/timeline';
