+++ project/zemn.me/next/pages/index.tsx
@@ -2,8 +2,8 @@ import Head from 'next/head';
 import * as bio from 'project/zemn.me/bio';
 import { TimeEye } from 'project/zemn.me/elements/TimeEye';
 import * as kenwood from 'project/zemn.me/next/assets/kenwood';
+import Timeline from 'project/zemn.me/next/components/timeline';
 import style from 'project/zemn.me/next/pages/index.module.css';
-import Timeline from 'project/zemn.me/next/pages/timeline';
 import * as lang from 'ts/react/lang';
 
 export default function Main() {
